// Projektovanje/frontend/blueStars-mobile/src/types/types.ts
export type LucideIconName = keyof typeof LucideIcons;
// ...
export interface MenuItemProps {
  id: string;
  icon: LucideIconName;
  text: string;
  onPressMenuItem: () => void;
}

export interface MenuSectionProps {
  title: string;
  items: MenuItemProps[];
}
// ...
export interface AppErrorDTO {
	note: string;
}